Object for managing running jobs on the redis queue

Object for managing running jobs on the redis queue

Object for managing running jobs on the redis queue

Public fields

controller

RRQ controller

Methods

Public methods

Queue$new()

Queue$submit()

Queue$number_of_workers()

Queue$get_status()

Queue$cancel()

Method new()

Create object, read configuration and setup Redis connection.

Usage

Queue$new(queue_id = NULL, logs_dir = "logs/worker")

Arguments

queue_id

ID of an existing queue to connect to, creates a new one if NULL (default NULL)

logs_dir

directory to store worker logs

Method submit()

Submit a job the Redis queue for runner to run.

Usage

Queue$submit(
  url,
  branch,
  ref,
  reportname,
  parameters,
  location,
  ssh_key = NULL
)

Arguments

url

The URL of the Git repository containing the reports.

branch

Name of git branch to checkout the repository.

ref

Git commit-ish value (e.g HEAD or commit hash or branch name).

reportname

Name of orderly report.

parameters

Parameters to run the report with.

location

Location of the outpack repository from which to pull dependencies and push the produced packet.

ssh_key

private SSH key for connecting to private repositories (Optional, default is NULL).

Method number_of_workers()

Gets the number of workers, only implemented until we add queue status for testing

Usage

Queue$number_of_workers()

Method get_status()

Gets statuses of packet runs in the queue.

Usage

Queue$get_status(task_ids, include_logs = TRUE)

Arguments

task_ids

Task ids to get status of.

include_logs

Whether to include logs in response or not

Returns

statuses of redis queue jobs

Method cancel()

Cancels a packet run

Usage

Queue$cancel(task_id)

Arguments

task_id

A single task to cancel

Returns

Nothing, or error if cancellation was not possible